Refactor memory access to use Int for addresses and unsigned types for values.
The `Memory` class methods (`readByte`, `readWord`, `readDWord`, `writeByte`, `writeWord`, `writeDWord`) now expect `Int` for addresses.
Values written to memory are now `UByte`, `UShort`, or `UInt`.
Values read from memory are converted to `Long` in the `Interpreter` for consistent handling, except for `readDWord` in `pop` which returns `Long` after being `UInt`.

The `printMemory` function in `Interpreter` now uses `Int` for physical addresses when reading bytes.

The `Parser` now uses `toInt()` for addresses and `toUByte()`, `toUShort()`, `toUInt()` for values when writing to memory for data directives.

A test value in `main.asm` was changed from `30Ah` to `3Ah`.

+++ src/main/kotlin/asm/Interpreter.kt
@@ -65,9 +65,9 @@ class Interpreter(
         }
         is Operand.RegOp -> cpu.readReg(op.reg) // CPU handles size for register reads
         is Operand.MemOp -> when (size) {
-            8 -> mem.readByte(addrOf(op)).toLong()
-            16 -> mem.readWord(addrOf(op)).toLong()
-            32 -> mem.readDWord(addrOf(op))
+            8 -> mem.readByte(addrOf(op).toInt()).toLong()
+            16 -> mem.readWord(addrOf(op).toInt()).toLong()
+            32 -> mem.readDWord(addrOf(op).toInt()).toLong()
             else -> error("Line $line: Unsupported size $size for memory read")
         }
         is Operand.LabelOp -> labels[op.name]?.let { it and 0xFFFFFFFFL } ?: error("Line $line: Unknown label: ${op.name}")
@@ -87,11 +87,11 @@ class Interpreter(
         when(op) {
             is Operand.RegOp -> cpu.writeReg(op.reg, value) // CPU handles size for register writes
             is Operand.MemOp -> {
-                val addr = addrOf(op)
+                val addr = addrOf(op).toInt()
                 when (size) {
-                    8 -> mem.writeByte(addr, value.toShort())
-                    16 -> mem.writeWord(addr, value.toInt())
-                    32 -> mem.writeDWord(addr, value)
+                    8 -> mem.writeByte(addr, value.toUByte())
+                    16 -> mem.writeWord(addr, value.toUShort())
+                    32 -> mem.writeDWord(addr, value.toUInt())
                     else -> error("Line $line: Unsupported size $size for memory write")
                 }
             }
@@ -200,7 +200,7 @@ class Interpreter(
 
         val ssBase = cpu.readReg(Reg.SS) // SS holds segment base
         val physicalStackAddress = (ssBase + newSpOffset) and 0xFFFFFFFFL
-        mem.writeDWord(physicalStackAddress, v) // Write Double Word (32-bit)
+        mem.writeDWord(physicalStackAddress.toInt(), v.toUInt()) // Write Double Word (32-bit)
     }
 
     /**
@@ -214,10 +214,10 @@ class Interpreter(
         val ssBase = cpu.readReg(Reg.SS) // SS holds segment base
         val physicalStackAddress = (ssBase + currentSpOffset) and 0xFFFFFFFFL
 
-        val v = mem.readDWord(physicalStackAddress) // Read Double Word (32-bit)
+        val v = mem.readDWord(physicalStackAddress.toInt()) // Read Double Word (32-bit)
         val newSpOffset = (currentSpOffset + 4) and 0xFFFFFFFFL // SP is 32-bit offset
         cpu.writeReg(Reg.ESP, newSpOffset)
-        return v
+        return v.toLong()
     }
 
     /**
@@ -419,7 +419,7 @@ class Interpreter(
             for (j in 0 until 16) {
                 val physicalAddress = i + j
                 if (physicalAddress < mem.size) { 
-                    val v = mem.readByte(physicalAddress.toLong())
+                    val v = mem.readByte(physicalAddress)
                     print("${v.toString(16).padStart(2, '0')} ")
                 } else {
                     print("   ")
